EXPERIMENTAL: added static database call / DatabaseFactory, rebuild NoteModel with static methods

+++ application/model/NoteModel.php
@@ -7,22 +7,15 @@
 class NoteModel
 {
     /**
-     * Constructor, expects a Database connection
-     * @param Database $database The Database object, from libs/Database.php
-     */
-    public function __construct(Database $database)
-    {
-        $this->database = $database;
-    }
-
-    /**
      * Get all notes (notes are just example data that the user has created)
      * @return array an array with several objects (the results)
      */
-    public function getAllNotes()
+    public static function getAllNotes()
     {
+        $database = DatabaseFactory::getFactory()->getConnection();
+
         $sql = "SELECT user_id, note_id, note_text FROM notes WHERE user_id = :user_id";
-        $query = $this->database->prepare($sql);
+        $query = $database->prepare($sql);
         $query->execute(array(':user_id' => Session::get('user_id')));
 
         // fetchAll() is the PDO method that gets all result rows
@@ -34,10 +27,12 @@ class NoteModel
      * @param int $note_id id of the specific note
      * @return object a single object (the result)
      */
-    public function getNote($note_id)
+    public static function getNote($note_id)
     {
+        $database = DatabaseFactory::getFactory()->getConnection();
+
         $sql = "SELECT user_id, note_id, note_text FROM notes WHERE user_id = :user_id AND note_id = :note_id LIMIT 1";
-        $query = $this->database->prepare($sql);
+        $query = $database->prepare($sql);
         $query->execute(array(':user_id' => Session::get('user_id'), ':note_id' => $note_id));
 
         // fetch() is the PDO method that gets a single result
@@ -49,10 +44,17 @@ class NoteModel
      * @param string $note_text note text that will be created
      * @return bool feedback (was the note created properly ?)
      */
-    public function createNote($note_text)
+    public static function createNote($note_text)
     {
+        if (!$note_text || strlen($note_text) == 0) {
+            Session::add('feedback_negative', FEEDBACK_NOTE_CREATION_FAILED);
+            return false;
+        }
+
+        $database = DatabaseFactory::getFactory()->getConnection();
+
         $sql = "INSERT INTO notes (note_text, user_id) VALUES (:note_text, :user_id)";
-        $query = $this->database->prepare($sql);
+        $query = $database->prepare($sql);
         $query->execute(array(':note_text' => $note_text, ':user_id' => Session::get('user_id')));
 
         if ($query->rowCount() == 1) {
@@ -70,10 +72,16 @@ class NoteModel
      * @param string $note_text new text of the specific note
      * @return bool feedback (was the update successful ?)
      */
-    public function updateNote($note_id, $note_text)
+    public static function updateNote($note_id, $note_text)
     {
+        if (!$note_id || !$note_text) {
+            return false;
+        }
+
+        $database = DatabaseFactory::getFactory()->getConnection();
+
         $sql = "UPDATE notes SET note_text = :note_text WHERE note_id = :note_id AND user_id = :user_id LIMIT 1";
-        $query = $this->database->prepare($sql);
+        $query = $database->prepare($sql);
         $query->execute(
             array(':note_id' => $note_id, ':note_text' => $note_text, ':user_id' => Session::get('user_id'))
         );
@@ -92,10 +100,16 @@ class NoteModel
      * @param int $note_id id of the note
      * @return bool feedback (was the note deleted properly ?)
      */
-    public function deleteNote($note_id)
+    public static function deleteNote($note_id)
     {
+        if (!$note_id) {
+            return false;
+        }
+
+        $database = DatabaseFactory::getFactory()->getConnection();
+
         $sql = "DELETE FROM notes WHERE note_id = :note_id AND user_id = :user_id LIMIT 1";
-        $query = $this->database->prepare($sql);
+        $query = $database->prepare($sql);
         $query->execute(array(':note_id' => $note_id, ':user_id' => Session::get('user_id')));
 
         if ($query->rowCount() == 1) {
